+++ include/simple_rover_locomotion/simple_rover_locomotion.hpp
@@ -10,7 +10,13 @@ class SimpleRoverLocomotion: public LocomotionMode
     SimpleRoverLocomotion(rclcpp::NodeOptions options, std::string node_name);
 
   private:
+  	// Position of the centre of rotation in x and y
+  	double centre_of_rotation_x;
+  	double centre_of_rotation_y;
+  	bool fully_steerable_;
+
     void rover_velocities_callback(const geometry_msgs::msg::Twist::SharedPtr msg);
+    bool check_steering_limitations();
     // bool enable();
 };